+++ openccjava/src/main/java/openccjava/ConversionPlanCache.java
@@ -172,9 +172,9 @@ public final class ConversionPlanCache {
             case S2TW: {
                 r1 = new ArrayList<>(Arrays.asList(d.st_phrases, d.st_characters));
                 if (punctuation) r1.add(d.st_punctuations);
-                r2 = Collections.singletonList(d.tw_variants);
+                r2 = Arrays.asList(d.tw_variants_phrases, d.tw_variants);
                 refs = new DictRefs(r1, unionCache.unionFor(punctuation ? UnionKey.S2T_PUNCT : UnionKey.S2T))
-                        .withRound2(r2, unionCache.unionFor(UnionKey.TwVariantsOnly));
+                        .withRound2(r2, unionCache.unionFor(UnionKey.TwVariantsPair));
                 break;
             }
             case TW2S: {
@@ -189,10 +189,10 @@ public final class ConversionPlanCache {
                 r1 = new ArrayList<>(Arrays.asList(d.st_phrases, d.st_characters));
                 if (punctuation) r1.add(d.st_punctuations);
                 r2 = Collections.singletonList(d.tw_phrases);
-                r3 = Collections.singletonList(d.tw_variants);
+                r3 = Arrays.asList(d.tw_variants_phrases, d.tw_variants);
                 refs = new DictRefs(r1, unionCache.unionFor(punctuation ? UnionKey.S2T_PUNCT : UnionKey.S2T))
                         .withRound2(r2, unionCache.unionFor(UnionKey.TwPhrasesOnly))
-                        .withRound3(r3, unionCache.unionFor(UnionKey.TwVariantsOnly));
+                        .withRound3(r3, unionCache.unionFor(UnionKey.TwVariantsPair));
                 break;
             }
             case TW2SP: {
@@ -206,9 +206,9 @@ public final class ConversionPlanCache {
             case S2HK: {
                 r1 = new ArrayList<>(Arrays.asList(d.st_phrases, d.st_characters));
                 if (punctuation) r1.add(d.st_punctuations);
-                r2 = Collections.singletonList(d.hk_variants);
+                r2 = Arrays.asList(d.hk_variants_phrases, d.hk_variants);
                 refs = new DictRefs(r1, unionCache.unionFor(punctuation ? UnionKey.S2T_PUNCT : UnionKey.S2T))
-                        .withRound2(r2, unionCache.unionFor(UnionKey.HkVariantsOnly));
+                        .withRound2(r2, unionCache.unionFor(UnionKey.HkVariantsPair));
                 break;
             }
             case HK2S: {
@@ -220,15 +220,15 @@ public final class ConversionPlanCache {
                 break;
             }
             case T2TW: {
-                r1 = Collections.singletonList(d.tw_variants);
-                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.TwVariantsOnly));
+                r1 = Arrays.asList(d.tw_variants_phrases, d.tw_variants);
+                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.TwVariantsPair));
                 break;
             }
             case T2TWP: {
                 r1 = Collections.singletonList(d.tw_phrases);
-                r2 = Collections.singletonList(d.tw_variants);
+                r2 = Arrays.asList(d.tw_variants_phrases, d.tw_variants);
                 refs = new DictRefs(r1, unionCache.unionFor(UnionKey.TwPhrasesOnly))
-                        .withRound2(r2, unionCache.unionFor(UnionKey.TwVariantsOnly));
+                        .withRound2(r2, unionCache.unionFor(UnionKey.TwVariantsPair));
                 break;
             }
             case TW2T: {
@@ -244,8 +244,8 @@ public final class ConversionPlanCache {
                 break;
             }
             case T2HK: {
-                r1 = Collections.singletonList(d.hk_variants);
-                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkVariantsOnly));
+                r1 = Arrays.asList(d.hk_variants_phrases, d.hk_variants);
+                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkVariantsPair));
                 break;
             }
             case HK2T: {