+++ openccjava/src/main/java/openccjava/UnionCache.java
@@ -44,9 +44,9 @@ final class UnionCache {
             case TwPhrasesOnly:
                 return getOrInit(slots.tw_phrases_only,
                         () -> StarterUnion.build(Collections.singletonList(d.tw_phrases)));
-            case TwVariantsOnly:
-                return getOrInit(slots.tw_variants_only,
-                        () -> StarterUnion.build(Collections.singletonList(d.tw_variants)));
+            case TwVariantsPair:
+                return getOrInit(slots.tw_variants_pair,
+                        () -> StarterUnion.build(Arrays.asList(d.tw_variants_phrases, d.tw_variants)));
             case TwPhrasesRevOnly:
                 return getOrInit(slots.tw_phrases_rev_only,
                         () -> StarterUnion.build(Collections.singletonList(d.tw_phrases_rev)));
@@ -56,9 +56,9 @@ final class UnionCache {
             case Tw2SpR1TwRevTriple:
                 return getOrInit(slots.tw2sp_r1_tw_rev_triple,
                         () -> StarterUnion.build(Arrays.asList(d.tw_phrases_rev, d.tw_variants_rev_phrases, d.tw_variants_rev)));
-            case HkVariantsOnly:
-                return getOrInit(slots.hk_variants_only,
-                        () -> StarterUnion.build(Collections.singletonList(d.hk_variants)));
+            case HkVariantsPair:
+                return getOrInit(slots.hk_variants_pair,
+                        () -> StarterUnion.build(Arrays.asList(d.hk_variants_phrases, d.hk_variants)));
             case HkRevPair:
                 return getOrInit(slots.hk_rev_pair,
                         () -> StarterUnion.build(Arrays.asList(d.hk_variants_rev_phrases, d.hk_variants_rev)));
@@ -91,11 +91,11 @@ final class UnionCache {
         final AtomicReference<StarterUnion> t2s = new AtomicReference<>();
         final AtomicReference<StarterUnion> t2s_punct = new AtomicReference<>();
         final AtomicReference<StarterUnion> tw_phrases_only = new AtomicReference<>();
-        final AtomicReference<StarterUnion> tw_variants_only = new AtomicReference<>();
+        final AtomicReference<StarterUnion> tw_variants_pair = new AtomicReference<>();
         final AtomicReference<StarterUnion> tw_phrases_rev_only = new AtomicReference<>();
         final AtomicReference<StarterUnion> tw_rev_pair = new AtomicReference<>();
         final AtomicReference<StarterUnion> tw2sp_r1_tw_rev_triple = new AtomicReference<>();
-        final AtomicReference<StarterUnion> hk_variants_only = new AtomicReference<>();
+        final AtomicReference<StarterUnion> hk_variants_pair = new AtomicReference<>();
         final AtomicReference<StarterUnion> hk_rev_pair = new AtomicReference<>();
         final AtomicReference<StarterUnion> jp_variants_only = new AtomicReference<>();
         final AtomicReference<StarterUnion> jp_rev_triple = new AtomicReference<>();